LayoutTests/imported/w3c/web-platform-tests/domxpath/resolver-non-string-result-expected.txt


PASS undefined
PASS null
PASS number
PASS boolean
FAIL symbol assert_throws_dom: function "() => {
    document.evaluate("/foo:bar", document.documentElement, resolver);
  }" threw object "TypeError: Cannot convert a symbol to a string" that is not a DOMException NAMESPACE_ERR: property "code" is equal to undefined, expected 14
FAIL object coercion (abrupt completion) assert_throws_dom: function "() => {
    document.evaluate("/foo:bar", document.documentElement, resolver);
  }" threw object "[object Object]" that is not a DOMException NAMESPACE_ERR: property "code" is equal to undefined, expected 14
